+++ src/bluetooth_adapters/const.py
@@ -10,4 +10,6 @@ FREEBSD_DEFAULT_BLUETOOTH_ADAPTER: Final = "ubt0hci"
 # Some operating systems hide the adapter address for privacy reasons (ex MacOS)
 DEFAULT_ADDRESS: Final = "00:00:00:00:00:00"
 
+EMPTY_MAC_ADDRESS: Final = "00:00:00:00:00:00"
+
 DEFAULT_CONNECTION_SLOTS: Final = 5